Restore the use of fixed names
Randomly generated names mean we will have to take into account the gender of those names. That seems like an additional complexity that I would not want to worry about, although it's not *too* hard...you just need to classify where a randomly generated name is male or female, and assign personal pronouns accordingly (or just assign it randomly if the name is ambigious).